+++ src/test/java/entities/OrderTest.java
@@ -9,7 +9,7 @@ import java.util.UUID;
 
 public class OrderTest {
     public static Order newOrder;
-    public static HashMap<Integer, Integer> orderQuantities = new HashMap<>();
+    public static HashMap<Long, Integer> orderQuantities = new HashMap<>();
     public static Date creationDate = new Date(2022, 10, 31, 0, 0);
 
     private UUID warehouseID;
@@ -18,8 +18,8 @@ public class OrderTest {
     @Before
     public void createTest(){
         // Puts some basic stuff in the order and builds it
-        orderQuantities.put(1, 10);
-        orderQuantities.put(2, 120);
+        orderQuantities.put(1L, 10);
+        orderQuantities.put(2L, 120);
 
         warehouseID = UUID.randomUUID();
         storeID = UUID.randomUUID();
@@ -36,8 +36,8 @@ public class OrderTest {
         Assertions.assertEquals(newOrder.getStatus(), Order.CREATED);
 
         // Tests the order quantities, ensures the keys are the same and actual quantities
-        HashMap<Integer, Integer> givenOrderQuantities = newOrder.getOrderQuantities();
-        for(int key: givenOrderQuantities.keySet()){
+        HashMap<Long, Integer> givenOrderQuantities = newOrder.getOrderQuantities();
+        for(long key: givenOrderQuantities.keySet()){
             Assertions.assertTrue(orderQuantities.containsKey(key));
             Assertions.assertEquals(orderQuantities.get(key), givenOrderQuantities.get(key));
         }
